+++ plugins/postcss-cascade-layers/src/compound-selector-order.ts
@@ -47,18 +47,18 @@ export function sortCompoundSelector(node) {
 
 	node.nodes.sort((a, b) => {
 		if (a.type === 'selector' && b.type === 'selector' && a.nodes.length && b.nodes.length) {
-			return selectorTypeOrder(a.nodes[0].value, a.nodes[0].type) - selectorTypeOrder(b.nodes[0].value, b.nodes[0].type);
+			return selectorTypeOrder(a.nodes[0], a.nodes[0].type) - selectorTypeOrder(b.nodes[0], b.nodes[0].type);
 		}
 
 		if (a.type === 'selector' && a.nodes.length) {
-			return selectorTypeOrder(a.nodes[0].value, a.nodes[0].type) - selectorTypeOrder(b.value, b.type);
+			return selectorTypeOrder(a.nodes[0], a.nodes[0].type) - selectorTypeOrder(b, b.type);
 		}
 
 		if (b.type === 'selector' && b.nodes.length) {
-			return selectorTypeOrder(a.value, a.type) - selectorTypeOrder(b.nodes[0].value, b.nodes[0].type);
+			return selectorTypeOrder(a, a.type) - selectorTypeOrder(b.nodes[0], b.nodes[0].type);
 		}
 
-		return selectorTypeOrder(a.value, a.type) - selectorTypeOrder(b.value, b.type);
+		return selectorTypeOrder(a, a.type) - selectorTypeOrder(b, b.type);
 	});
 }
 
@@ -67,9 +67,6 @@ function selectorTypeOrder(selector, type) {
 		return selectorTypeOrderIndex.pseudoElement;
 	}
 
-	if (type === 'pseudo' && selector && selector.indexOf('::') === 0) {
-		return selectorTypeOrderIndex.pseudoElement;
-	}
 	return selectorTypeOrderIndex[type];
 }