+++ plugins/postcss-pseudo-class-any-link/src/compound-selector-order.js
@@ -47,25 +47,26 @@ export function sortCompoundSelector(node) {
 
 	node.nodes.sort((a, b) => {
 		if (a.type === 'selector' && b.type === 'selector' && a.nodes.length && b.nodes.length) {
-			return selectorTypeOrder(a.nodes[0].value, a.nodes[0].type) - selectorTypeOrder(b.nodes[0].value, b.nodes[0].type);
+			return selectorTypeOrder(a.nodes[0], a.nodes[0].type) - selectorTypeOrder(b.nodes[0], b.nodes[0].type);
 		}
 
 		if (a.type === 'selector' && a.nodes.length) {
-			return selectorTypeOrder(a.nodes[0].value, a.nodes[0].type) - selectorTypeOrder(b.value, b.type);
+			return selectorTypeOrder(a.nodes[0], a.nodes[0].type) - selectorTypeOrder(b, b.type);
 		}
 
 		if (b.type === 'selector' && b.nodes.length) {
-			return selectorTypeOrder(a.value, a.type) - selectorTypeOrder(b.nodes[0].value, b.nodes[0].type);
+			return selectorTypeOrder(a, a.type) - selectorTypeOrder(b.nodes[0], b.nodes[0].type);
 		}
 
-		return selectorTypeOrder(a.value, a.type) - selectorTypeOrder(b.value, b.type);
+		return selectorTypeOrder(a, a.type) - selectorTypeOrder(b, b.type);
 	});
 }
 
 function selectorTypeOrder(selector, type) {
-	if (type === 'pseudo' && selector && selector.indexOf('::') === 0) {
-		return selectorTypeOrderIndex['pseudoElement'];
+	if (parser.isPseudoElement(selector)) {
+		return selectorTypeOrderIndex.pseudoElement;
 	}
+
 	return selectorTypeOrderIndex[type];
 }